Vulkan: Split off QueueVk
